Move basal results to Calculated Basal sub-page, matching Calculated ISF
The main Autotune screen now shows only toggles + run controls. The basal
profile table, carb/sensitivity summary, delete, and save-to-pump button all
live in a new Calculated Basal sub-page, accessible via a NavigationLink that
shows the last-run timestamp as a subtitle.

Both sub-pages (Calculated Basal, Calculated ISF) follow the same pattern:
the link is always present when Use Autotune is on, with a "No data yet"
subtitle until the first run completes.
